remove unneeded files
