+++ src/rngToolsUi/gen3/battleVideo/battleVideo.tsx
@@ -12,8 +12,6 @@ import {
   FormFieldTable,
   FormikSelect,
   Field,
-  Button,
-  Flex,
 } from "~/components";
 import { formatLargeInteger } from "~/utils/formatLargeInteger";
 import InstructionsNoBattleNoExisting from "./instructions_no_battle_no_existing.mdx";
@@ -29,8 +27,6 @@ import {
   gen3ConsoleOptions,
   gen3Consoles,
 } from "~/types/console";
-import { Wild3Action } from "~/rngTools";
-import { isFishingAction } from "../wild/utils";
 
 // No waiting in battle
 const ADV_MISC_NO_BATTLE = 100; // Advances before last input not caused by frames when not waiting in battle
@@ -106,7 +102,7 @@ const createInitialValues = (fixedData?: FixedData): FormState => ({
   targetAdvance: fixedData?.targetAdvance ?? 50_000,
   isUpdatingExisting: fixedData?.isUpdatingExisting ?? false,
   console: fixedData?.consoleType ?? "GBA",
-  forFishing: isFishingAction(fixedData?.action ?? "SweetScentLand"),
+  forFishing: false,
   considerWaitingInBattle: true,
   displayAdvancedBreakdown: false,
   useRecommendedBuffer: true,
@@ -131,14 +127,12 @@ const MyFields = ({
     name: "targetAdvance",
   });
 
-  const existingBattleVideoAdv = fixedData?.existingBattleVideoAdv ?? 0;
-
   const fields: Field[] = [
     {
       label: fixedData?.isAfterPainting
         ? "Existing Battle Video advances after painting"
         : "Existing Battle Video advances",
-      input: "~" + formatLargeInteger(existingBattleVideoAdv),
+      input: formatLargeInteger(fixedData?.existingBattleVideoAdv ?? 0),
       show: fixedData?.isUpdatingExisting === true,
     },
     {
@@ -150,11 +144,7 @@ const MyFields = ({
       label: fixedData?.isAfterPainting
         ? "Target advance after painting"
         : "Target advance",
-      input:
-        formatLargeInteger(targetAdvance) +
-        (fixedData?.isUpdatingExisting
-          ? ` (+${formatLargeInteger(targetAdvance - existingBattleVideoAdv)})`
-          : ``),
+      input: formatLargeInteger(targetAdvance),
       show: fixedData != null,
     },
     {
@@ -188,7 +178,7 @@ const MyFields = ({
       label: "Leave sufficient buffer for fishing?",
       input: <FormikSwitch<FormState> name="forFishing" />,
       indent: 1,
-      show: useRecommendedBuffer && fixedData?.action == null,
+      show: useRecommendedBuffer,
     },
     {
       label: "Buffer (advance)",
@@ -335,12 +325,11 @@ const calculateWithoutBattle = (opts: FormState) => {
     ],
     battleVideoInfo: {
       initialAdv,
-      battleVideoAdv: initialAdv + targetAdvAtVideo,
+      battleVideoAdv: targetAdvAtVideo,
       safetyBufferAdv,
-      targetAdv: initialAdv + targetAdvance,
+      targetAdv: targetAdvance,
       withBattle: false,
       isUpdatingExisting: opts.isUpdatingExisting,
-      consoleType: opts.console,
     },
   };
 };
@@ -504,7 +493,6 @@ const calculateWithBattle = (opts: FormState) => {
       targetAdv: opts.targetAdvance,
       withBattle: true,
       isUpdatingExisting: opts.isUpdatingExisting,
-      consoleType: opts.console,
     },
   };
 };
@@ -529,27 +517,13 @@ type FixedData = {
   existingBattleVideoAdv: number;
   isAfterPainting: boolean;
   consoleType?: Gen3Console;
-  action?: Wild3Action;
-};
-
-export type BattleVideoInfo = {
-  targetPaintingAdvs: {
-    before: number;
-    after: number;
-  };
-  battleVideoAdvAfterPainting: number;
-  consoleType: Gen3Console | null;
 };
 
 export type Props = {
   fixedData?: FixedData;
-  setBattleVideoAdv?: (
-    adv: number | null,
-    consoleType: Gen3Console | null,
-  ) => void;
 };
 
-export const BattleVideo = ({ fixedData, setBattleVideoAdv }: Props) => {
+export const BattleVideo = ({ fixedData }: Props) => {
   const [milliseconds, setMilliseconds] = React.useState<number[]>([]);
   const [timerLabels, setTimerLabels] = React.useState<string[]>([]);
   const [submitError, setSubmitError] = React.useState("");
@@ -565,7 +539,6 @@ export const BattleVideo = ({ fixedData, setBattleVideoAdv }: Props) => {
     targetAdv: number;
     withBattle: boolean;
     isUpdatingExisting: boolean;
-    consoleType: Gen3Console | null;
   } | null>(null);
 
   const onSubmit: RngToolSubmit<FormState> = async (opts) => {
@@ -575,7 +548,7 @@ export const BattleVideo = ({ fixedData, setBattleVideoAdv }: Props) => {
       timerLabels,
       breakdown,
       battleVideoInfo,
-    } = calculate(opts);
+    } = await calculate(opts);
     setSubmitError(submitError);
     setMilliseconds(milliseconds);
     setTimerLabels(timerLabels);
@@ -660,42 +633,6 @@ export const BattleVideo = ({ fixedData, setBattleVideoAdv }: Props) => {
           startButtonTrackerId="battle_video_timer_start"
           stopButtonTrackerId="battle_video_timer_stop"
         />
-      )}
-
-      {setBattleVideoAdv != null && (
-        <Flex gap={10} mt={20}>
-          <div>Step outcome: </div>
-          {(battleVideoInfo != null || initialValues.isUpdatingExisting) && (
-            <Button
-              trackerId="wild3_battle_video_created"
-              onClick={() => {
-                if (battleVideoInfo != null) {
-                  setBattleVideoAdv(
-                    battleVideoInfo.battleVideoAdv,
-                    battleVideoInfo.consoleType,
-                  );
-                } else {
-                  setBattleVideoAdv(initialValues.existingBattleVideoAdv, null);
-                }
-              }}
-            >
-              Battle Video was created at advance{" ~"}
-              {formatLargeInteger(
-                battleVideoInfo?.battleVideoAdv ??
-                  initialValues.existingBattleVideoAdv,
-              )}
-            </Button>
-          )}
-
-          <Button
-            trackerId="wild3_battle_video_skip"
-            onClick={() => {
-              setBattleVideoAdv(null, null);
-            }}
-          >
-            Skip (no Battle Video created)
-          </Button>
-        </Flex>
       )}
     </>
   );
