--- src/rngToolsUi/gen3/wild/instructions_calib_skip_setup.mdx
@@ -1,6 +1,0 @@
-The following section assumes that you already know the setup for your target Pokémon. If not, go back to the first step.
-
-1. Specify the setup of your Target Pokémon.
-1. Press "Calculate Target".
-1. Verify that the calculated target match what you really want.
-1. Specify Battle Video information.
