--- src/rngToolsUi/gen3/wild/instructions_calib_without_battle_video.mdx
@@ -1,15 +1,0 @@
-1. In-game, save at the map of the setup, standing on top of a patch of grass, water, or in front of a smashable rock.
-1. On the website, start the timers.
-1. When the 1st timer reaches 0, press `START + SELECT + A + B` to soft reset the game.
-1. Select your save file.
-1. Prepare the encounter:
-   - For Sweet Scent encounters, quickly open your Pokémon party. Move your marker on `SWEET SCENT` of a Pokémon, and wait.
-   - For fishing encounters, use your fishing rod until you have a Pokémon on the hook. Wait with the message "A Pokémon's on the hook!" displayed.
-     - The RNG manipulation can be successful even if you fail hooking a Pokémon the first try.
-   - For Rock Smash encounters, interact with the smashable rock. Wait with the message "[Pokémon] used ROCK SMASH!" displayed.
-1. When the 2nd timer reaches 0, press `A` to trigger the encounter.
-1. Catch the Pokémon, even if it's not your target Pokémon. It will be used to calibrate.
-1. Open the Pokémon summary and fill the fields on the website with the Pokémon stats. Press "Generate".
-1. If there are no results or only results with Low confidence ratings, make sure the fields are correct.
-1. Click on the "Update Calibrate" button of the result with the highest confidence rating (the first result). This will update the timers to improve your odds of catching your target Pokémon.
-1. Repeat all the previous steps until you catch your target Pokémon.
