+++ src/rngToolsUi/gen3/wild/wild3Calib.tsx
@@ -1,103 +1,46 @@
-import React from "react";
-import { Flex, MultiTimer, Field, Input, Select, Button } from "~/components";
+import React, { useState } from "react";
+import { Flex, MultiTimer, Field, Input, Select } from "~/components";
 import { FormFieldTable } from "~/components/formFieldTable";
-import {
-  TargetSetup,
-  Wild3CalibTargetSetupInput,
-} from "./wild3CalibTargetSetupInput";
-import { Wild3CalibCaughtMon } from "./wild3CalibCaughtMon.component";
+import { FormState as TargetSetup, Wild3CalibTarget } from "./wild3CalibTarget";
+import { Wild3CalibCaughtMon } from "./wild3CalibCaughtMon";
 import {
   Gen3Console,
   gen3ConsoleFpsMap,
   gen3ConsoleOptions,
 } from "~/types/console";
-import {
-  formatLargeInteger,
-  formatLargeIntegerWithSign,
-} from "~/utils/formatLargeInteger";
-import { formatActionName, formatLeadName, formatMapName } from "./utils";
-import { BattleVideoInfo } from "../battleVideo/battleVideo";
-import { AllOrNone } from "~/types";
-import { BattleVideoInfoInput } from "./wild3CalibBattleVideoInfoInput";
-import { calculateTargetSetupResult } from "./calculateTargetSetupResult";
-import { formatHex } from "~/utils/formatHex";
 
-import Instructions_calib_skip_setup from "./instructions_calib_skip_setup.mdx";
-import Instructions_calib_with_battle_video from "./instructions_calib_with_battle_video.mdx";
-import Instructions_calib_without_battle_video from "./instructions_calib_without_battle_video.mdx";
-
-type Props = AllOrNone<{
-  targetSetup: TargetSetup;
-  battleVideoInfo: BattleVideoInfo;
-  clearAll?: () => void;
-  displayInstructions?: boolean;
-}>;
-
-export const Wild3Calib = ({
-  targetSetup: targetSetupProp,
-  battleVideoInfo: battleVideoInfoProp,
-  clearAll,
-  displayInstructions = true,
-}: Props) => {
+export const Wild3Calib = () => {
   const [targetSetup, setTargetSetup] = React.useState<TargetSetup | null>(
-    targetSetupProp ?? null,
+    null,
   );
 
-  const [targetSetupResult, setTargetSetupResult] =
-    React.useState<React.ReactNode>(null);
+  const [consoleType, setConsoleType] = useState<Gen3Console>("GBA");
 
-  React.useEffect(() => {
-    if (targetSetup == null) {
-      return setTargetSetupResult(null);
-    }
-
-    calculateTargetSetupResult(targetSetup).then(setTargetSetupResult);
-  }, [targetSetup]);
-
-  const [battleVideoInfo, setBattleVideoInfo] =
-    React.useState<BattleVideoInfo | null>(battleVideoInfoProp ?? null);
-
-  const [consoleTypeFromInput, setConsoleTypeFromInput] =
-    React.useState<Gen3Console>("GBA");
+  /** calibration is always for target advance after painting. 
+      calibration is not used if the painting seed is not confirmed */
   const [calibrationAndOffset, setCalibrationAndOffset] = React.useState(0);
 
-  const targetSetupInputForm = () => (
-    <Flex vertical gap={10}>
-      {displayInstructions && <Instructions_calib_skip_setup />}
-      <Wild3CalibTargetSetupInput setTargetSetup={setTargetSetup} />
-      {targetSetupResult != null && (
-        <FormFieldTable
-          fields={[
-            {
-              label: "Target Pokémon",
-              input: targetSetupResult,
-            },
-          ]}
-        />
-      )}
-    </Flex>
-  );
+  const calibrationIsActive =
+    targetSetup !== null &&
+    (!targetSetup.usingPaintingReseeding ||
+      targetSetup.isPaintingSeedConfirmed);
 
-  if (targetSetup == null) {
-    return targetSetupInputForm();
-  }
+  const setLatestHitAdv = calibrationIsActive
+    ? (hitAdv: {
+        frame_before_painting: number;
+        adv_after_painting: number;
+      }) => {
+        setCalibrationAndOffset(
+          calibrationAndOffset + hitAdv.adv_after_painting - targetForTimer,
+        );
+      }
+    : undefined;
 
-  const setLatestHitAdv = (hitAdv: {
-    frame_before_painting: number;
-    adv_after_painting: number;
-  }) => {
-    setCalibrationAndOffset(
-      calibrationAndOffset + hitAdv.adv_after_painting - targetForTimer,
-    );
-  };
+  const targetForTimer = targetSetup?.targetAdvance ?? 0;
 
-  const targetForTimer = targetSetup?.targetPaintingAdvs.after ?? 0;
-
-  const initialAdv = battleVideoInfo?.battleVideoAdvAfterPainting ?? 0;
+  const initialAdv = targetSetup?.existingBattleVideoAdv ?? 0;
 
   const advFromTimer = targetForTimer - initialAdv - calibrationAndOffset;
-
-  const consoleType = battleVideoInfo?.consoleType ?? consoleTypeFromInput;
 
   const milliseconds = [
     5000,
@@ -108,176 +51,60 @@ export const Wild3Calib = ({
     "Trigger Sweet Scent",
   ];
 
-  const battleVideoInfoInputForm = () => {
-    if (targetSetup == null) {
-      return null;
-    }
-
-    return (
-      <BattleVideoInfoInput
-        targetPaintingAdvs={targetSetup.targetPaintingAdvs}
-        setBattleVideoInfo={setBattleVideoInfo}
-      />
-    );
-  };
-
-  const consoleProp = battleVideoInfoProp?.consoleType;
-  const calibFields: Field[] = [
-    // consoleType has 3 possible sources:
-    //  - battleVideoInfoProp.consoleType
-    //  - battleVideoInfoInput (if battleVideoInfoProp == null)
-    //  - here if (battleVideoInfoProp != null but battleVideoInfoProp.consoleType is null)
-
-    // Don't show Console field if battleVideoInfoProp == null because it's already shown in battleVideoInfoInput
-    {
-      label: "Console",
-      show: battleVideoInfoProp != null,
-      input:
-        battleVideoInfoProp != null &&
-        battleVideoInfoProp.consoleType == null ? (
-          <Select<Gen3Console>
-            name="console"
-            value={consoleTypeFromInput}
-            options={gen3ConsoleOptions}
-            onSelect={(val) => {
-              setConsoleTypeFromInput(val);
-            }}
-          />
-        ) : (
-          (gen3ConsoleOptions.find((opt) => opt.value === consoleProp)?.label ??
-          "")
-        ),
-    },
-    {
-      label: "Calibration + Offset (advance)",
-      input: (
-        <Input
-          name="offset"
-          onChange={(event) => {
-            const num = Number(event.target.value);
-            setCalibrationAndOffset(Number.isFinite(num) ? num : 0);
-          }}
-          value={calibrationAndOffset}
-        />
-      ),
-    },
-  ];
-
-  const canDoCalib =
-    battleVideoInfo != null || targetSetup.targetPaintingAdvs.before === 0;
-
-  const infoFromPrevSteps = () => {
-    if (targetSetupProp == null) {
-      return null;
-    }
-
-    const usingPaintingReseeding =
-      battleVideoInfoProp.targetPaintingAdvs.before > 0;
-
-    const fields: Field[] = [
-      {
-        label: "Map",
-        input: formatMapName(targetSetupProp.map),
-      },
-      {
-        label: "Player action",
-        input: formatActionName(targetSetupProp.action),
-      },
-      {
-        label: "Lead",
-        input: formatLeadName(targetSetupProp.lead),
-      },
-      {
-        label: "Lead Cycle Speed",
-        input: targetSetupProp.usingAverageLeadCycleSpeed
-          ? "Average"
-          : targetSetupProp.leadCycleSpeed,
-        show: targetSetupProp.lead !== "Egg",
-      },
-      {
-        label: "Target Method",
-        input: targetSetupProp.targetMethod,
-      },
-      {
-        label: "Target frame before painting",
-        input: `${formatLargeInteger(battleVideoInfoProp.targetPaintingAdvs.before)} (Seed: ${formatHex(battleVideoInfoProp.targetPaintingAdvs.before, 2)})`,
-        show: usingPaintingReseeding,
-      },
-      {
-        label: "Battle Video advance",
-        input: formatLargeInteger(
-          battleVideoInfoProp.battleVideoAdvAfterPainting,
-        ),
-        show: battleVideoInfoProp.battleVideoAdvAfterPainting > 0,
-      },
-      {
-        label: "Target advance",
-        input:
-          formatLargeInteger(targetSetup.targetPaintingAdvs.after) +
-          (initialAdv > 0
-            ? ` (${formatLargeIntegerWithSign(targetSetup.targetPaintingAdvs.after - initialAdv)} from Battle Video)`
-            : ``),
-        show: !usingPaintingReseeding,
-      },
-      {
-        label: "Target advance after painting",
-        input: `${formatLargeInteger(targetSetup.targetPaintingAdvs.after)} (${formatLargeIntegerWithSign(targetSetup.targetPaintingAdvs.after - initialAdv)} from Battle Video)`,
-        show: usingPaintingReseeding,
-      },
-      {
-        label: "Target Pokémon",
-        input: targetSetupResult,
-      },
-    ];
-
-    return (
-      <Flex vertical>
-        <h3>Info from previous steps</h3>
-        <Flex ml={20} vertical>
-          <FormFieldTable fields={fields} />
-          {clearAll != null && (
-            <Button
-              trackerId="wild3Calib_clearAll"
-              danger
-              maxWidth={150}
-              size="small"
-              onClick={clearAll}
-            >
-              Clear All
-            </Button>
-          )}
-        </Flex>
-      </Flex>
-    );
-  };
-
-  const inputForms = () => {
-    return (
-      <>
-        {targetSetupInputForm()}
-        {battleVideoInfoInputForm()}
-      </>
-    );
-  };
-
-  const usingBattleVideo =
-    (battleVideoInfo?.battleVideoAdvAfterPainting ?? 0) > 0;
+  const fields: Field[] =
+    targetSetup != null
+      ? [
+          {
+            label: "Target advance",
+            input: <>{targetSetup.targetAdvance}</>,
+            show: !targetSetup.usingPaintingReseeding,
+          },
+          {
+            label: "Target frame before painting",
+            input: <>{targetSetup.targetFrameBeforePainting}</>,
+            show: targetSetup.usingPaintingReseeding,
+          },
+          {
+            label: "Target advance after painting",
+            input: <>{targetSetup.targetAdvance}</>,
+            show: targetSetup.usingPaintingReseeding,
+          },
+          {
+            label: "Calibration + Offset (advance)",
+            input: (
+              <Input
+                name="offset"
+                onChange={(event) => {
+                  const num = Number(event.target.value);
+                  setCalibrationAndOffset(Number.isFinite(num) ? num : 0);
+                }}
+                value={calibrationAndOffset}
+              />
+            ),
+          },
+          {
+            label: "Console",
+            input: (
+              <Select<Gen3Console>
+                name="console"
+                value={consoleType}
+                options={gen3ConsoleOptions}
+                onSelect={(val) => {
+                  setConsoleType(val);
+                }}
+              />
+            ),
+          },
+        ]
+      : [];
 
   return (
     <Flex gap={32} vertical>
-      {targetSetupProp == null ? inputForms() : infoFromPrevSteps()}
+      <Wild3CalibTarget setTargetSetup={setTargetSetup} />
 
-      {canDoCalib && (
+      {targetSetup !== null && (
         <>
-          <FormFieldTable fields={calibFields} />
-
-          {displayInstructions &&
-            (usingBattleVideo ? (
-              <Instructions_calib_with_battle_video />
-            ) : (
-              <Instructions_calib_without_battle_video />
-            ))}
-
+          <FormFieldTable fields={fields} />
           <MultiTimer
             milliseconds={milliseconds}
             labels={labels}
